route organized by creating different files for routing and helper functions
