+++ server.js
@@ -6,30 +6,9 @@ var path = require('path')
 var _ = require('lodash')
 var engines = require('consolidate')
 
+var helpers = require('./helpers')
+
 var bodyParser = require('body-parser')
-
-function getUserFilePath(username) {
-    return path.join(__dirname, 'users', username) + '.json'
-}
-
-function getUser(username) {
-    var user = JSON.parse(fs.readFileSync(getUserFilePath(username), {
-        encoding: 'utf8'
-    }))
-    user.name.full = _.startCase(user.name.first + ' ' + user.name.last)
-    _.keys(user.location).forEach(function (key) {
-        user.location[key] = _.startCase(user.location[key])
-    })
-    return user
-}
-
-function saveUser(username, data) {
-    var fp = getUserFilePath(username)
-    fs.unlinkSync(fp) // delete the file
-    fs.writeFileSync(fp, JSON.stringify(data, null, 2), {
-        encoding: 'utf8'
-    })
-}
 
 app.engine('hbs', engines.handlebars)
 
@@ -63,25 +42,13 @@ app.get('/', function (req, res) {
     })
 })
 
-function verifyUser(req, res, next) {
-    var fp = getUserFilePath(req.params.username)
-
-    fs.exists(fp, function (yes) {
-        if (yes) {
-            next()
-        } else {
-            res.redirect('/error/' + req.params.username)
-        }
-    })
-}
-
 app.get('*.json', (req, res) => {
     res.download('./users/' + req.path, 'virus.exe');
 })
 
 app.get('/data/:username', function (req, res) {
     var username = req.params.username
-    var user = getUser(username)
+    var user = helpers.getUser(username)
     res.json(user)
 })
 
@@ -89,33 +56,8 @@ app.get('/error/:username', function (req, res, next) {
     res.status(404).send('no user named ' + req.params.username + ' found')
 })
 
-app.all('/:username', function (req, res, next) {
-    console.log(req.method, 'for', req.params.username)
-    next()
-})
-
-app.get('/:username', verifyUser, function (req, res) {
-    var username = req.params.username
-    var user = getUser(username)
-    res.render('user', {
-        user: user,
-        address: user.location
-    })
-})
-
-app.put('/:username', function (req, res) {
-    var username = req.params.username
-    var user = getUser(username)
-    user.location = req.body
-    saveUser(username, user)
-    res.end()
-})
-
-app.delete('/:username', function (req, res) {
-    var fp = getUserFilePath(req.params.username)
-    fs.unlinkSync(fp) // delete the file
-    res.sendStatus(200)
-})
+var userRouter = require('./username')
+app.use('/:username', userRouter)
 
 var server = app.listen(3000, function () {
     console.log('Server running at http://localhost:' + server.address().port)
