+++ views/user.hbs
@@ -61,7 +61,7 @@
     <div class="pic">
         <img src="/profilepics/{{user.username}}_med.jpg">
         <p>
-            <a href="#" onclick="edit()">Edit</a>
+            <a href="/{{user.username}}/edit" onclick="edit()">Edit</a>
             <a href="#" onclick="del()">Delete</a>
         </p>
     </div>
